+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - database: Fixed crash when schema generation should fail with `schema_modified`.
 - hasura: Stop using deprecated schema/metadata API.
 - index: Fixed unnecessary prefetching of migration originations in `operation` index.
+- index: Remove disabled index from the dispatcher queue.
 - sentry: Flush and reopen session daily.
 - tzkt: Fixed `OperationData.type` field value for migration originations.
 - tzkt: Added missing `last_level` argument to migration origination fetching methods.